pyproject.toml
[tool.poetry]
name = "vahid-bot"
version = "2.0.0"
description = "General management bot for the BUCS Discord server"
authors = ["Alicja Mahr <braxton@mahr.cc>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "vahid_bot"}]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.0.0"
coloredlogs = "^15.0.1"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
